fix unit tests

+++ src/Seedwork/Infra.Crosscutting/Collections/IPagedCollection.cs
@@ -5,5 +5,5 @@ namespace Goal.Seedwork.Infra.Crosscutting.Collections;
 public interface IPagedCollection : IEnumerable
 {
     int PageCount { get; }
-    int TotalCount { get; }
+    long TotalCount { get; }
 }
